+++ requests/serve_contracts.go
@@ -33,11 +33,33 @@ func GetGeneralContractRoutes() []string {
 	return routes
 }
 
+func GetGeneralContractDataFiber(ctx *fiber.Ctx) error {
+
+	// assemble key from route
+	var key string
+	route := strings.Split(ctx.Route().Path, `/`)
+
+	for index, part := range route {
+		if index > 1 {
+			key += ":" + part
+		} else if index == 1 {
+			key += part
+		}
+	}
+
+	rdb := config.RDB
+	val, err := rdb.Get(context.Background(), key).Result()
+	if err != nil {
+		panic(err)
+	}
+	return ctx.SendString(val)
+}
+
 // Processed Pairs
 func QueryPairs(ctx *fiber.Ctx) error {
-	val, err := getStoreValueFromKey(rediskeys.ProcessedPairs)
+	val, err := GetStoreValueFromKey(rediskeys.ProcessedPairs)
 	if err != nil {
-		return redisKeyNotFound(ctx, rediskeys.ProcessedPairs)
+		return RedisKeyNotFound(ctx, rediskeys.ProcessedPairs)
 	}
 	return ctx.Status(StatusOkay).SendString(val)
 }
